requirements.txt
python-binance
TA-lib
numpy
flask